fix(#2482): re-arrange common.ts and react specific exports
