+++ libs/react-components/src/common/types.ts
@@ -1,0 +1,8 @@
+import { Spacing } from "@abgov/ui-components-common";
+
+export interface Margins {
+  mt?: Spacing;
+  mr?: Spacing;
+  mb?: Spacing;
+  ml?: Spacing;
+}
