+++ libs/react-components/src/experimental/form/task-list.tsx
@@ -1,5 +1,5 @@
 import { ReactNode } from "react";
-import { Margins } from "@abgov/ui-components-common";
+import { Margins } from "../../common/types";
 
 interface WCProps extends Margins {
   heading: string;
