+++ libs/react-components/src/index.ts
@@ -59,6 +59,7 @@ export * from "./lib/three-column-layout/three-column-layout";
 export * from "./lib/tooltip/tooltip";
 export * from "./lib/two-column-layout/two-column-layout";
 export * from "./lib/filter-chip/filter-chip";
+export * from "./common/types";
 
 export type {
   GoabSpinnerType,
@@ -109,25 +110,18 @@ export type {
   GoabTextAreaCountBy,
   GoabTextAreaOnChangeDetail,
   GoabTextAreaOnKeyPressDetail,
-  GoabTabsProps,
   GoabTabsOnChangeDetail,
   GoabTableVariant,
   GoabTableSortDirection,
-  GoabTableProps,
   GoabTableOnSortDetail,
   GoabSpacerHorizontalSpacing,
   GoabSpacerVerticalSpacing,
-  GoabSpacerProps,
-  GoabSkeletonProps,
   GoabSkeletonType,
   GoabSkeletonSize,
   GoabRadioGroupOrientation,
-  GoabRadioGroupProps,
-  GoabRadioItemProps,
   GoabCircularProgressVariant,
   GoabCircularProgressSize,
   GoabPopoverPosition,
-  GoabPopoverProps,
   GoabNotificationType,
   GoabAriaLiveType,
   GoabServiceLevel,
@@ -142,9 +136,6 @@ export type {
   OnFocus,
   OnBlur,
   OnKeyPress,
-  GoabInputProps,
-  GoabNumberInputProps,
-  GoabDateInputProps,
   GoabIconFilledType,
   GoabIconType,
   GoabIconSize,
@@ -154,7 +145,6 @@ export type {
   NumericSpacing,
   TShirtSpacing,
   Spacing,
-  Margins,
   GoabBlockDirection,
   GoabBlockAlignment,
   GoabLinkButtonType,
