+++ libs/react-components/src/lib/callout/callout.tsx
@@ -1,4 +1,5 @@
-import { GoabCalloutAriaLive, GoabCalloutSize, GoabCalloutType, GoabCalloutIconTheme, Margins } from "@abgov/ui-components-common";
+import { GoabCalloutAriaLive, GoabCalloutSize, GoabCalloutType, GoabCalloutIconTheme } from "@abgov/ui-components-common";
+import { Margins } from "../../common/types";
 
 interface WCProps extends Margins {
   heading?: string;
