+++ libs/react-components/src/lib/container/container.tsx
@@ -3,8 +3,9 @@ import {
   GoabContainerPadding,
   GoabContainerType,
   GoabContainerWidth,
-  Margins,
+  
 } from "@abgov/ui-components-common";
+import { Margins } from "../../common/types";
 import { ReactNode } from "react";
 
 interface WCProps extends Margins {
