+++ libs/react-components/src/lib/date-picker/date-picker.tsx
@@ -1,5 +1,6 @@
 import { useEffect, useRef } from "react";
-import { GoabDatePickerOnChangeDetail, Margins } from "@abgov/ui-components-common";
+import { GoabDatePickerOnChangeDetail } from "@abgov/ui-components-common";
+import { Margins } from "../../common/types";
 
 interface WCProps extends Margins {
   ref: React.RefObject<HTMLElement>;
