+++ libs/react-components/src/lib/dropdown/dropdown.tsx
@@ -1,8 +1,8 @@
 import {
   GoabDropdownOnChangeDetail,
-  GoabIconType,
-  Margins,
+  GoabIconType
 } from "@abgov/ui-components-common";
+import { Margins } from "../../common/types";
 import { useEffect, useRef } from "react";
 
 interface WCProps extends Margins {
