+++ libs/react-components/src/lib/form-item/form-item.tsx
@@ -1,8 +1,9 @@
 import {
   GoabFormItemLabelSize,
   GoabFormItemRequirement,
-  Margins,
+  
 } from "@abgov/ui-components-common";
+import { Margins } from "../../common/types";
 
 interface WCProps extends Margins {
   label?: string;
