+++ libs/react-components/src/lib/form-stepper/form-stepper.tsx
@@ -1,5 +1,6 @@
-import { GoabFormStepperOnChangeDetail, Margins } from "@abgov/ui-components-common";
+import { GoabFormStepperOnChangeDetail } from "@abgov/ui-components-common";
 import { ReactNode, useEffect, useRef } from "react";
+import { Margins } from "../../common/types";
 
 interface WCProps extends Margins {
   ref?: React.MutableRefObject<HTMLElement | null>;
