+++ libs/react-components/src/lib/grid/grid.tsx
@@ -1,4 +1,5 @@
-import { Margins, Spacing } from "@abgov/ui-components-common";
+import { Spacing } from "@abgov/ui-components-common";
+import { Margins } from "../../common/types";
 
 interface WCProps extends Margins {
   gap?: Spacing;
