+++ libs/react-components/src/lib/pages/pages.tsx
@@ -1,4 +1,4 @@
-import { Margins } from "@abgov/ui-components-common";
+import { Margins } from "../../common/types";
 import { ReactNode } from "react";
 
 interface WCProps extends Margins {
