+++ libs/react-components/src/lib/side-menu-group/side-menu-group.tsx
@@ -1,6 +1,6 @@
 import { ReactNode } from "react";
-import { GoabIconType, Margins } from "@abgov/ui-components-common";
-
+import { GoabIconType,  } from "@abgov/ui-components-common";
+import { Margins } from "../../common/types";
 interface WCProps extends Margins {
   heading: string;
   icon?: GoabIconType;
